add little more stability to unit tests

+++ Example/CedarUnitTests/Specs/DTCollectionViewController+VerifyItem.m
@@ -44,7 +44,7 @@
             [self raiseInvalidSectionException];
         }
     }
-    int itemsInSection = [self.collectionView numberOfItemsInSection:sectionNumber];
+    NSInteger itemsInSection = [self.collectionView numberOfItemsInSection:sectionNumber];
     if (itemsInSection!=[section count])
     {
         [self raiseInvalidSectionException];
